==> 02_AlterDataset.py <==
# -*- coding: utf-8 -*-
"""
Input - S3Dataset.csv
1. Optional - Remove images from 'Bad' cameras
2. Add following columns
- name - description - type - default value
- annotaion - contains annotation json - json  - empty
- threadNum - related thread number - int - Nan
- isAnnotated - was this image annotated - bool - False
Output - Dataset.csv
"""

BadViewCameraIDs = ['C0150201','C0150202','C0150204','C0150209','C0152600','C0162802','C0163201','C0163209','C0165100','C0252701','C0255000','C0351800','C0352408','C0354202','C0356105','C0350802','C0367502','C0450903','C0456500','C0650408','C0651808','C0855706','C0855707','C0855708','C0858706','C0858708','C0951208','C0951801','C0951802','C0951809','C0952001','C0952002','C0952009','C1050601','C1050602','C1050607','C1050608','C1055309','C1250907','C1254802','C1255007','C1259903','C1259904','C1259905','C1259906','C1259907','C1259908','C1260108','C1450301','C1450302','C1450309','C1453308','C1453508','C1454007','C1457305','C1457808']

CloseViewCameraIDs = ['C0150309','C0150409','C0150809','C0150909','C0151009','C0151109','C0152109','C0152209','C0152309','C0152409','C0153209','C0153509','C0153609','C0153809',
'C0158309','C0158409','C0162209','C0162409','C0162509','C0162609','C0162709','C0167309','C0167709','C0170309','C0250701','C0250709','C0250809','C0251109',
'C0251209','C0251504','C0251509','C0251709','C0252008','C0252009','C0252109','C0252209','C0252309','C0252409','C0252509','C0252609','C0252709','C0252809',
'C0254509','C0254609','C0255109','C0256309','C0256909','C0350409','C0350509','C0350609','C0350709','C0351909','C0352209','C0352309','C0352409','C0352509',
'C0352909','C0353109','C0353209','C0353309','C0353409','C0353809','C0353909','C0354009','C0354309','C0354409','C0354809','C0355209','C0355309','C0355709',
'C0355809','C0355909','C0356009','C0356109','C0356209','C0356309','C0356409','C0356509','C0356609','C0356709','C0356809','C0357009','C0357109','C0357209',
'C0357309','C0357509','C0357609','C0357809','C0357909','C0358009','C0358109','C0366109','C0366309','C0366509','C0450209','C0450309','C0450409','C0450609',
'C0450709','C0450809','C0450909','C0451309','C0451809','C0451909','C0452209','C0452409','C0452709','C0452809','C0453009','C0453109','C0453309','C0453408',
'C0453409','C0453509','C0453609','C0453809','C0454009','C0454109','C0454209','C0454309','C0454409','C0454709','C0454809','C0456609','C0456701','C0456702',
'C0456709','C0461209','C0461309','C0461409','C0461509','C0461609','C0650109','C0650309','C0650409','C0650509','C0650808','C0650809','C0650909','C0651009',
'C0651109','C0651209','C0651309','C0651409','C0651609','C0651709','C0651809','C0850509','C0850609','C0850809','C0850909','C0851109','C0851709','C0851809',
'C0851909','C0852009','C0852109','C0852209','C0852409','C0852609','C0852709','C0852809','C0852909','C0853009','C0853109','C0853209','C0853309','C0853409',
'C0853509','C0853609','C0853709','C0853809','C0854109','C0854309','C0854409','C0854509','C0854709','C0854909','C0855209','C0855309','C0855609','C0855709',
'C0855809','C0855909','C0856009','C0856109','C0857202','C0857209','C0857309','C0857409','C0857509','C0857609','C0858309','C0858409','C0858509','C0858609',
'C0858709','C0858809','C0950109','C0950209','C0950409','C0950509','C0950609','C0950709','C0950909','C0951109','C0951209','C0951309','C0951409','C0951509',
'C0951609','C0951909','C0952209','C0952309','C0952409','C0952509','C0953109','C1050109','C1050209','C1050309','C1050409','C1050509','C1050609','C1050709',
'C1050809','C1051109','C1051209','C1051309','C1051409','C1051509','C1051609','C1051709','C1051809','C1051909','C1052009','C1052109','C1052209','C1052409',
'C1052509','C1052609','C1052809','C1052909','C1053009','C1053109','C1053209','C1053309','C1053409','C1053509','C1053609','C1053709','C1053809','C1053909',
'C1054109','C1054209','C1054309','C1054609','C1054709','C1054909','C1055009','C1055109','C1055209','C1055309','C1250109','C1250209','C1250309','C1250509',
'C1250609','C1250709','C1250809','C1250909','C1251009','C1251109','C1251409','C1251509','C1251609','C1251709','C1251909','C1252009','C1252109','C1252209',
'C1252409','C1252609','C1252709','C1252809','C1252909','C1253009','C1253109','C1253209','C1253309','C1253409','C1253609','C1253709','C1253809','C1253909',
'C1254109','C1254209','C1254309','C1254509','C1254709','C1254809','C1255009','C1255109','C1255209','C1255409','C1255509','C1255709','C1255809','C1256009',
'C1256109','C1256309','C1256409','C1256509','C1256609','C1256709','C1256809','C1256909','C1257009','C1257109','C1257309','C1258109','C1258209','C1258409',
'C1258609','C1258709','C1258809','C1258909','C1259009','C1259109','C1259809','C1259903','C1259904','C1259905','C1259906','C1259907','C1259908','C1259909',
'C1260109','C1260209','C1260409','C1260509','C1261809','C1450409','C1450509','C1450709','C1450809','C1451009','C1451209','C1451409','C1451509','C1451609',
'C1451709','C1451909','C1452009','C1452209','C1452309','C1452409','C1452509','C1452609','C1452709','C1452809','C1453009','C1453109','C1453209','C1453409',
'C1453509','C1453609','C1453709','C1453809','C1453909','C1454009','C1454109','C1454209','C1454309','C1454409','C1454609','C1455409','C1455509','C1455609',
'C1455709','C1455809','C1455909','C1456009','C1456109','C1456409','C1456609','C1457109','C1457309','C1457409','C1457509','C1457809','C1458109','C1458209',
'C1458309','C1458409','C1458509','C1458609','C0352709']

UnitedBadCamerasIDs = BadViewCameraIDs + CloseViewCameraIDs

from tqdm import tqdm
import pandas as pd
import numpy as np
import os

SAVE_FOLDER = 'AnnotationResults'
REMOVE_BAD_IMAGES = True

df = pd.read_csv(os.path.join(SAVE_FOLDER,'S3Dataset.csv'))

if (REMOVE_BAD_IMAGES)
    for camID in tqdm(UnitedBadCamerasIDs):
        del_filter = df['image_url'].str.contains(pat = camID)
        df.drop(index=df[del_filter].index, inplace=True)

df['annotation'] = ''
df['threadNum'] = np.nan
df['isAnnotated'] = False

df.to_csv(os.path.join(SAVE_FOLDER,'Dataset.csv'),index=False)
